+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.45.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.45.1...@codecademy/gamut-illustrations@0.45.2) (2024-05-29)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
 ### [0.45.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.45.0...@codecademy/gamut-illustrations@0.45.1) (2024-05-01)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations