+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.410](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.409...@codecademy/gamut-kit@0.6.410) (2024-05-30)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.409](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.408...@codecademy/gamut-kit@0.6.409) (2024-05-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
 ### [0.6.408](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.407...@codecademy/gamut-kit@0.6.408) (2024-05-22)
 
 **Note:** Version bump only for package @codecademy/gamut-kit