+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.3.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@16.2.1...@codecademy/gamut-styles@16.3.0) (2024-05-29)
+
+### Features
+
+- **Form + GridForm:** Add actual required optional + required form text ([e712b44](https://github.com/Codecademy/gamut/commit/e712b4476554a7816ba556fe5482ad0451ebde33))
+
 ### [16.2.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@16.2.0...@codecademy/gamut-styles@16.2.1) (2024-05-01)
 
 ### Performance Improvements